// src/enums/route-ids-enum.ts
export enum RouteIdsEnum {
  HOME = 'HOME',
  EMPLOYEES = 'EMPLOYEES',
  ATTENDANCE_LOGS = 'ATTENDANCE_LOGS',
  NATIONALITIES = 'NATIONALITIES',
  CITIES = 'CITIES',
  REGIONS = 'REGIONS',
  PERMISSION_REASONS = 'PERMISSION_REASONS',
  NOTIFICATION_CHANNELS = 'NOTIFICATION_CHANNELS',
  DASHBOARD = 'DASHBOARD',
  HOLIDAYS = 'HOLIDAYS',
}